Move `LoadEcdsa...` into `pkg/cosign/keys.go`
Chains is consuming this public function, but because `./cmd/cosign/cli` pulls in `glog` we hit the `-log_dir` problem.  With this change it can drop it's dependency on `cli`.
